requirement.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crcmod==1.7
cryptography==42.0.8
Django==5.0.7
django-aliyun-oss2-storage==0.1.5
idna==3.7
jmespath==0.10.0
oss2==2.18.6
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
